services/vision-expert/src/vision_expert/expert.py
"""Vision Expert: Image analysis with CLIP.

Architecture:
1. Image input → Load and preprocess
2. CLIP encoding → Get image and text embeddings
3. Similarity computation → Match image to text labels
4. Classification → Return top-k predictions
5. Human review for critical decisions
"""

from __future__ import annotations

import base64
import io
import logging
from typing import Any

from PIL import Image

logger = logging.getLogger(__name__)


class VisionExpert:
    """Vision Expert with CLIP for image understanding.

    Uses OpenAI's CLIP (Contrastive Language-Image Pre-training) for:
    - Zero-shot image classification
    - Image-text similarity
    - Visual question answering
    - Object detection (with additional models)
    """

    def __init__(
        self,
        model_name: str = "openai/clip-vit-base-patch32",
        device: str = "cpu",
    ):
        """Initialize Vision Expert.

        Args:
            model_name: CLIP model name
            device: Device for inference (cpu/cuda)
        """
        self.model_name = model_name
        self.device = device

        # Initialize CLIP (placeholder - in production load transformers model)
        logger.info(f"Vision Expert initialized with model: {model_name}")
        logger.warning("Using placeholder image analysis (load CLIP in production)")

    async def process(self, request: dict[str, Any]) -> dict[str, Any]:
        """Process expert request (main entry point).

        Args:
            request: Expert request with capability and parameters

        Returns:
            Expert response with recommendation and metadata

        Example:
            >>> response = await expert.process({
            ...     "capability": "image_classification",
            ...     "query": "Classify this image",
            ...     "context": {
            ...         "image_url": "https://example.com/image.jpg",
            ...         "labels": ["cat", "dog", "bird"]
            ...     }
            ... })
        """
        capability = request.get("capability")
        query = request.get("query", "")
        context = request.get("context", {})

        logger.info(f"Processing Vision Expert request: capability={capability}")

        if capability == "image_classification":
            return await self._classify_image(query, context)
        elif capability == "image_text_similarity":
            return await self._image_text_similarity(query, context)
        elif capability == "visual_qa":
            return await self._visual_qa(query, context)
        elif capability == "object_detection":
            return await self._detect_objects(query, context)
        else:
            return {
                "recommendation": f"Unknown capability: {capability}",
                "confidence": 0.0,
                "reasoning": "Unknown capability",
                "metadata": {"error": "unknown_capability"},
                "requires_approval": True,
            }

    async def _classify_image(self, query: str, context: dict) -> dict[str, Any]:
        """Classify image into categories.

        Args:
            query: Classification query
            context: Context with image and labels

        Returns:
            Classification results with confidence scores
        """
        image_data = context.get("image_data")  # Base64 encoded
        image_url = context.get("image_url")
        labels = context.get("labels", [])

        if not labels:
            return {
                "recommendation": "No classification labels provided",
                "confidence": 0.0,
                "reasoning": "Cannot classify without target labels",
                "metadata": {"error": "missing_labels"},
                "requires_approval": True,
            }

        # Load image (placeholder)
        image = self._load_image(image_data, image_url)

        # Classify with CLIP (placeholder)
        predictions = self._placeholder_classify(image, labels)

        # Determine confidence level
        top_score = predictions[0]["score"]
        confidence = top_score

        reasoning = f"Image classified with top prediction '{predictions[0]['label']}' (confidence: {top_score:.2f})"

        return {
            "recommendation": f"Classification: {predictions[0]['label']}",
            "confidence": confidence,
            "reasoning": reasoning,
            "metadata": {
                "predictions": predictions,
                "top_k": 3,
                "model": self.model_name,
            },
            "requires_approval": confidence < 0.8,  # Require approval for low confidence
        }

    async def _image_text_similarity(self, query: str, context: dict) -> dict[str, Any]:
        """Compute similarity between image and text descriptions.

        Args:
            query: Text query
            context: Context with image

        Returns:
            Similarity scores
        """
        image_data = context.get("image_data")
        image_url = context.get("image_url")
        texts = context.get("texts", [query])

        # Load image
        image = self._load_image(image_data, image_url)

        # Compute similarities (placeholder)
        similarities = self._placeholder_similarity(image, texts)

        top_match = similarities[0]

        return {
            "recommendation": f"Best match: '{top_match['text']}' (similarity: {top_match['score']:.2f})",
            "confidence": top_match["score"],
            "reasoning": "Computed CLIP image-text similarity scores",
            "metadata": {
                "similarities": similarities,
                "query": query,
            },
            "requires_approval": False,  # Similarity search doesn't need approval
        }

    async def _visual_qa(self, query: str, context: dict) -> dict[str, Any]:
        """Answer questions about images.

        Args:
            query: Question about the image
            context: Context with image

        Returns:
            Answer to visual question
        """
        image_data = context.get("image_data")
        image_url = context.get("image_url")

        # Load image
        image = self._load_image(image_data, image_url)

        # Generate answer (placeholder - in production use VQA model)
        answer = self._placeholder_vqa(image, query)

        return {
            "recommendation": answer,
            "confidence": 0.7,
            "reasoning": "Visual question answered using CLIP + reasoning. Verify accuracy.",
            "metadata": {
                "question": query,
                "model": "CLIP (placeholder VQA)",
            },
            "requires_approval": True,  # VQA requires human verification
        }

    async def _detect_objects(self, query: str, context: dict) -> dict[str, Any]:
        """Detect objects in image.

        Args:
            query: Detection query
            context: Context with image

        Returns:
            Detected objects with bounding boxes
        """
        image_data = context.get("image_data")
        image_url = context.get("image_url")

        # Load image
        image = self._load_image(image_data, image_url)

        # Detect objects (placeholder - in production use YOLO/Detectron2)
        detections = self._placeholder_detect(image)

        return {
            "recommendation": f"Detected {len(detections)} objects",
            "confidence": 0.75,
            "reasoning": "Object detection completed (requires specialized model in production)",
            "metadata": {
                "detections": detections,
                "model": "Placeholder (use YOLO/Detectron2)",
            },
            "requires_approval": True,  # Object detection requires verification
        }

    def _load_image(self, image_data: str | None = None, image_url: str | None = None) -> Image.Image | None:
        """Load image from base64 data or URL.

        Args:
            image_data: Base64 encoded image
            image_url: URL to image

        Returns:
            PIL Image or None
        """
        if image_data:
            try:
                # Decode base64
                image_bytes = base64.b64decode(image_data)
                return Image.open(io.BytesIO(image_bytes))
            except Exception as e:
                logger.error(f"Failed to decode image data: {e}")
                return None

        if image_url:
            # In production, fetch from URL
            logger.warning("Image URL loading not implemented (placeholder)")
            return None

        return None

    def _placeholder_classify(self, image: Image.Image | None, labels: list[str]) -> list[dict[str, Any]]:
        """Placeholder classification (replace with CLIP in production).

        Args:
            image: PIL Image
            labels: Classification labels

        Returns:
            Predictions with scores
        """
        # Simple mock predictions
        import random

        random.seed(42)  # Deterministic for testing

        predictions = []
        for label in labels:
            score = random.uniform(0.1, 0.95)
            predictions.append({"label": label, "score": score})

        # Sort by score descending
        predictions.sort(key=lambda x: x["score"], reverse=True)

        return predictions[:5]  # Top 5

    def _placeholder_similarity(self, image: Image.Image | None, texts: list[str]) -> list[dict[str, Any]]:
        """Placeholder similarity computation.

        Args:
            image: PIL Image
            texts: Text descriptions

        Returns:
            Similarity scores
        """
        import random

        random.seed(42)

        similarities = []
        for text in texts:
            score = random.uniform(0.3, 0.95)
            similarities.append({"text": text, "score": score})

        similarities.sort(key=lambda x: x["score"], reverse=True)

        return similarities

    def _placeholder_vqa(self, image: Image.Image | None, question: str) -> str:
        """Placeholder VQA (replace with VQA model in production).

        Args:
            image: PIL Image
            question: Question about image

        Returns:
            Answer
        """
        # Simple template-based answers
        question_lower = question.lower()

        if "color" in question_lower:
            return "The dominant colors appear to be blue and white."
        elif "how many" in question_lower:
            return "I can see approximately 3-5 objects, but verification recommended."
        elif "what is" in question_lower or "what are" in question_lower:
            return "The image appears to contain common objects. Specific identification requires human verification."
        else:
            return "Answer requires visual analysis. Please review the image for accurate response."

    def _placeholder_detect(self, image: Image.Image | None) -> list[dict[str, Any]]:
        """Placeholder object detection.

        Args:
            image: PIL Image

        Returns:
            Detected objects
        """
        # Mock detections
        return [
            {
                "label": "object_1",
                "confidence": 0.92,
                "bbox": {"x": 10, "y": 20, "width": 100, "height": 150},
            },
            {
                "label": "object_2",
                "confidence": 0.85,
                "bbox": {"x": 200, "y": 50, "width": 80, "height": 120},
            },
        ]
