services/vision-expert/pyproject.toml
[project]
name = "vision-expert"
version = "0.1.0"
description = "Vision Expert with CLIP for image analysis for OpenDataGov"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "OpenDataGov Team"}
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pillow>=10.1.0",
    "httpx>=0.25.0",
    "odg-core>=0.1.0",
    "pydantic>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.282",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vision_expert"]
